Chore: 상단 nav 중복 구현 정리
