+++ app/(cardThema)/me/(cards)/layout.tsx
@@ -1,8 +1,7 @@
 'use client';
 
 import { usePathname } from "next/navigation";
-import { HeaderWithBefore } from "@/features/auth/ui/header-with-before";
-import { AuthHeader } from "@/features/auth/ui/auth-header";
+import { AuthHeader, HeaderWithBefore } from "@/features/auth/ui/auth-header";
 import { Container } from "@/shared/ui/container";
 
 export default function CardsLayout({
